now also handling text data types with custom length
mysql's documentation mentions that integer data type width is to be
treated at application level [1] and as such, mysql only throws a warning
about it if it exceeds the width in the table definition.

the situation is different in the case of text data types [2] where the
data is truncated if it does not fit.

this commit adds logic to handle the text data types with custom length.

[1] http://dev.mysql.com/doc/refman/5.7/en/numeric-type-attributes.html
[2] http://dev.mysql.com/doc/refman/5.7/en/blob.html