Replacing bcdiv with floatval
Removing arbitrary precision arithmetic in favor of using
floating-point arithmetic instead. There's in fact two problems we're trying to solve:
a) we're adding the dependency php-bcmath which apparently doesn't even ship with PHP
b) using bcdiv in certain edge-cases throws an exception.
   example:

   php7.2 -r '$a=1;$b=18446744073709551615; $b1=-$b; bcdiv($a,$b,4);'
   PHP Warning:  bcdiv(): Division by zero in Command line code on line 1

So apparently, using bcdiv does not solve the problem that it was trying to solve.

Instead, @no-gravity suggests we switch to floating point numbers.
According to [2] [3], it looks like the range we can use now is
[2.2250738585072E-308,1.7976931348623E+308]

[1] https://github.com/wsdookadr/fieldtop/pull/5#discussion-diff-71966834
[2] https://externals.io/message/105037#105045
[3] https://twiserandom.com/php/data-types/what-is-a-float-or-a-double-in-php/

Note:
There's one TODO left here. The .travis.yml file needs fixing because
phpunit seems to always exit with an exit code 0 and reports that all
tests have passed (even though they fail locally). I don't know how
to fix that part at the moment, so I'm leaving it like that for now, until
there's more time to look at it.